requirements.txt
pandas==2.2.3
openai==1.78.1
python-dotenv==1.1.0
pytest==8.3.4
pytest-mock==3.14.0